[PFD2-Javadocs]
If a servlet calls request.getSession() true with no current session, and
the response has already been committed, and we are using cookies for
session id tracking, throw IllegalStateException.  This is because the
headers and cookies have already been sent, so the client would never find
out about the newly created session.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285943 13f79535-47bb-0310-9956-ffa450edef68