Same changes as in runtime 11 - the code is almost identical with the one in tomcat-4, with Constants and JasperException moved in the same package and all external dependencyes removed ( well, replaced with a Liaiason ).
The code is very similar with runtime 1.1, with few diffs because of the new
features, minus performance optimizations.

It should be possible ( even quite easy ) to use the same code base for
both - with probably one or 2 "ifs" to provide the right behavior.
( not a priority for now, but when we start 'real' optimizations of the
runtime )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283487 13f79535-47bb-0310-9956-ffa450edef68