Make it possible to dynamically select the Mapper implementation class to use for the StandardEngine implementation.  This is done by setting the "mapperClass" attribute of the <Engine> element.
Add a new Mapper implementation for StandardEngine that is optimized for
performance.  Among other things, it caches mappings for all define host names
and aliases, and remembers each mapping to the default host (if any) after
the first time.  No object creations occur during request processing.

The default mapper is still the previous standard one.  If you wish to
experiment, select the new implementation by the following in your
server.xml file:

    <Engine ... mapperClass="org.apache.catalina.core.FastEngineMapper" .../>

When we are satisfied that this class works correctly, it will become the
default.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@285974 13f79535-47bb-0310-9956-ffa450edef68