The generators.
Most of the files can be easily translated to a xsl template, and the
code is quite clean.

We need first to integrate 4.0, and get everything to work. Plus finish the
changes in the support classes ( mangler, jspservlet, etc), then
convert the parser to SAX. Converting the generators to XSLT templates
is probably the last step.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283534 13f79535-47bb-0310-9956-ffa450edef68