Introduce the concept of a ContainerServlet, which require access to Catalina internal implementation objects to perform their functions.  Such servlets must be: - In a subpackage of "org.apache.catalina" - Loaded by the Catalina class loader (rather than the web app class   loader)
Modify the only two servlets that require this functionality
(InvokerServlet and ManagerServlet) to gain access to Catalina internals
via the wrapper that is assigned by the container, rather than by casting
the servlet API objects that are passed.  This allows these servlets to be
oblivious to the introduction of facades around the servlet API objects.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286060 13f79535-47bb-0310-9956-ffa450edef68