Begin converting the Catalina developer docs to XML format, so that they can be generated by whatever Anakia-based or XSLT-based generation script is selected.
To begin the process of documenting current behavior of Tomcat 4.0
components (and to provide folks trying to write tests some assertions
about what correct behavior is), I've provided "Functional Specifications"
that document the current behavior of the default file-serving servlet and
the invoker servlet.  Folks working on various portions of Tomcat are
encouraged to do the same thing for those portions.  This will help us
build better tests, and also serve as useful information for newcomers to
the Tomcat source base to understand what is going on.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286085 13f79535-47bb-0310-9956-ffa450edef68