Added ( refactored ) UTF encoder and decoder.
The code used to be part of Byte/Char Chunk, but had many bugs and it was hard
to optimize.

Note that we don't implement M$ encoding scheme ( which is not standard and
may cause many problems ), but it could be implemented.

There is still work to be done for decoding char[] - the result of the
conversion is byte, and it has to be converted ( somehow ) to char, but
you can't do that without a b->c converter.

( this will happen for RequestDispatchers for example - a workaround is to
not encode "extended" chars )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283593 13f79535-47bb-0310-9956-ffa450edef68