Few big changes here...
The "new" code is now in use ( has been here for few months, but we kept
using the old code for safety ). The most important thing is that
now we corectly decode ( if charset is provided ) - the old code didn't.

There is still some old code in use ( only used by RequestDispatcher), but
the common code is now fixed.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283601 13f79535-47bb-0310-9956-ffa450edef68