For easy compilation - added the binaries jasper34 depends on.
The tomcat_util.jar is also used ( in part ) by the connector. In time
we'll migrate it to commons ( or keep it in jakarta-tomcat-connectors ).
As we are refactoring the code in jasper it will be replaced with
pluggable interfaces - but I don't think it's a priority for now.

We can assume we'll use some generic utils ( at least the pools ) - and
importing the code in jasper will not help.

tomcat-core is imported to build the 3.3 specific module - will go away
after we improve the build.xml.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304738 13f79535-47bb-0310-9956-ffa450edef68