Few more file moves.
The old mangling scheme is specific to the old servlet, no need to keep it
in the generator.

I expect a full rewrite of the servlet, and maybe use of standard URLClasLoader
and normal file names ( I think Glenn mentioned it's already done in 4.0,
so we just need to merge the code ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304746 13f79535-47bb-0310-9956-ffa450edef68