Some changes in runtime to simplify the generated code.
If we extend HttpJspBase ( and our code depends on jasper.runtime.HttpJspBase
anyway ) we can avoid some complexity in the generated code. If we extend
any other class we need of course to generate the old jspService method.

I don't know if using more runtime internals in the generated code is good
or bad, but I know it's almost impossible to generate code that will not
depend on some jasper runtime ( tag pool, HttpJspBase, etc ).

Since we have to use the runtime, it's better to do it in a clean way.

Note that this doesn't affect jasper portability or code independence -
the runtime can/must be included with the generated code ( if jspc is used)
or must be available on the container ( if the container is using jasper )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304778 13f79535-47bb-0310-9956-ffa450edef68