Few fixes I missed last time.
We need to get "out" at the begining, since the page could call
pushBody and we'll get a different out.

Another small fix - out could be BodyContentImpl ( and I have no idea
how to flush or if it should be flushed - or what happens if some
pushBody() is called without pop ).

( one watchdog test does that ).

Keeping the old behavior ( i.e. do nothing ) is probably the best idea.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304782 13f79535-47bb-0310-9956-ffa450edef68