Moved the web.xml and tld reading out of generator and into liaison.
The container could pass the information from it's reading of web.xml
and may have better/faster ways of reading the xml.

We provide a default ( the old DOM based code ), treat it as a black box
( and start a new impl. if you want new features ).

Some other changes - the API allows now to do the reading only once
( but the container liaison will have to keep the TagLibraries instance
around - right now we create/garbage collect a lot )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304783 13f79535-47bb-0310-9956-ffa450edef68