Added 2 more hooks in the generator.
The first one allows generators to declare dependencies. The liaison could
use those to implement reloading ( including the famous "include" bug ).

The second one allows static content generators to pre-declare the
chunks. It is not used right now, but will work in association with
the RuntimeLiaison to optimize the whole output system ( pre-conversion
to byte[], caching of static content on the server side, etc ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304786 13f79535-47bb-0310-9956-ffa450edef68