A bit of cleanup and reorganization.
EmbededTomcat is the main entry point and starter for tomcat.

Tomcat.java is a nice program that can execute a number of tasks, process
the command line, etc. It delegates to EmbededTomcat whatever is related
with starting tomcat, and to other objects to stop, etc.

Main is a java wrapper that sets the class paths for clean separation
and executes Tomcat.

With a bit of time, we should refactor Main in a general-purpose
application starter ( in tomcat/util ), and play a bit more with
command line processing and introspection ( maybe with a simple
properties file that would drive Main ). Tomcat.java can disapear if
Main can take over the dispatching and command line processing.

The original idea was to use patterns like in "ant" for Main - only
with command line options instead of XML attributes. A services-like
properties file could allow Main to work with any application following
the setXXX and execute() patterns.

No time for that right now - but if anyone wants to try I would be happy to
help.


PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283658 13f79535-47bb-0310-9956-ffa450edef68