Various changes to support the line mapping.
Javagen adds a getter for the line number.

Various generators fixed to not include \n in the generated code
( that could be also fixed by treating all println as printMultiLn, but
it's better to have generators respect the conventions.

( we have a stack trace each time we detect a generator including \n, for
debugging - it'll be removed after we fix everything )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304795 13f79535-47bb-0310-9956-ffa450edef68