Initial code to use the new line map.
We don't do anything special - this is just the original code that does
the mapping.

Very inefficient ( 2 Writers, Readers, many Srings, linear search ) - it can
be changed later ( possibly in a general purpose util ). This is called only
on error - shouldn't be a frequent operation.

The code should do a println with the detected location in the JSP file

( small bug - the line is 0-based, need to increment it )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304797 13f79535-47bb-0310-9956-ffa450edef68