Minor update to previous commit.  The previous commit contains a change I had planned to commit separately.  This change was to port an improvement found in the JspInterceptor.
The improvement is to avoid violating the "post data must remain available"
requirement found in section 5.1 of the servlet 2.2 spec when trying to
determine if "jsp_precompile" is specified.  Parameters are obtained from the
request only if "jsp_precompile" is found in the query string.  Also, the query
parameters will not be logged if "jsp_precompile" is not found and the request
method is "POST".

This means that putting a "jsp_precompile" parameter in POST data will no
longer work.  I'd don't believe this violates the intent of the spec with respect
to "jsp_precompile" nor is likely to cause any hardship.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283665 13f79535-47bb-0310-9956-ffa450edef68