Removed the mangler and options from tomcat33 liaison.
We're moving ( as proposed by Mel ) to a toolbox aproach, with a number
of manglers, options, etc that you can combine to implement the adapter
between jasper and a container.

JspServlet will remain the "default" one for easiest integration (
after refactoring, loader changes and some extra docs ), but to provide
container-specific features to jasper you'll need to use callbacks
and maybe modified implementations for the tools.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304808 13f79535-47bb-0310-9956-ffa450edef68