Same kind of changes to tomcat33 adapter.
A lot of code still need to be moved either in the base class for
JavaCompiler ( detection, etc) or in toolbox.

Note: there is a small bug in JspInterceptor reloading introduced by the
refactoring, I'm working on it ( the deps system need a bit of abstraction
in order to allow the same toolbox aproach and integrate with Glenn's one
class loader per jsp ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304817 13f79535-47bb-0310-9956-ffa450edef68