Compiler is the main helper to convert jsp to java and to class.
This fix eliminates a lot of duplicated code from various liaisons, moves logging
and all common operations here and cleans up ( a bit ) the process.

The compile() method was removed - it didn't provide enough control on error
reporting and guessing.

To compile java files you need to create a JavaCompiler and use it - then we
can access the message ( and process it to extract line numbers and map them ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304820 13f79535-47bb-0310-9956-ffa450edef68