Eliminate unused code ( duplicate that was consolidated, etc ).
Move code from JspInterceptor to liaison ( to simplify interceptor and to
allow the move to a "toolkit" pattern )

Fix the method signatures.

Fix in JspInterceptor to allow the use of a JspServlet without any setting changes
in the module ( for example if a webapp is using a foreign jsp impl )
XXX this might need to be ported back to 33.

In JspInterceptor, eliminate JasperLiaison ( it's not needed, many of the ideas are
migrated into Compiler )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304828 13f79535-47bb-0310-9956-ffa450edef68