The first serious cleanup for JspServlet.
I tested it and it seems to work fine ( assuming you get the right classpath :-).
I'll send a mail with the details on using JspServlet, it needs jasper.jar
and all required deps in the classpath ( not available in tomcat33, since the
container is isolated - so you can either put it in common or in the webapp ).
There are few issues regarding the use of JspServlet, the code is still far behind
the interceptor - but could be used for simple things.

The good news is that most of the spaghetti that ended up in JspServlet is
gone, I hope most people will be able to understand the code now ( I had big problems
with that ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@304831 13f79535-47bb-0310-9956-ffa450edef68