Few fixes related with decoding 8859_1.
Replaced default encoding from UTF8 to 8859_1. While I think it would be
better to default to UTF8 in low-level utils like MessageBytes, fact is that
most of the uses for MB is in servlet container where 8859_1 is required
 as default impl. That's pretty bad as most RFCs and recent standards seem to be
moving to UTF.

( we could make it non-final, but than it would be even worse ).

The solution is to make sure the encoding is propagated and set by the caller
( we could throw an exception - just to make sure the problem is handled )

Also removed an extra log.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@295466 13f79535-47bb-0310-9956-ffa450edef68