Add a call to clear any application-originated context attributes during a restart, in order to avoid dangling references to object instances created by the old class loader (after a reload, you would get a ClassCastException trying to access such an attribute).
HOWEVER:  The call to clearAttributes() is currently commented out,
because it causes ClassNotFoundexceptions trying to restart listeners and
filters, for a reason I cannot fathom yet, on the test suite additions to
be checked in next.

There is also an additional problem -- we're not re-init()-ing the load on
startup servlets after a restart.  That will be fixed once we figure out
what's going on with clearing attributes.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286193 13f79535-47bb-0310-9956-ffa450edef68