Merged ( copied ) the code from Ajp13 into Ajp14, removed "extend Ajp13".
The main reason is that we need to keep Ajp13 stable while we want to
do some refactoring in Ajp14 - separate the marshaling, add an extension
mechanism to be used to implement and add new callbacks, drop Ajp14Packet
and start using the ByteChunk.

In other words the architecture will change from "Ajp14 extends Ajp13" into
a AjpMarshaling ( which will work for both ), and AjpCallback ( with
"original" 13 callbacks plus new ajp14 callbacks ), i.e. composition instead of
inheritance.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@295506 13f79535-47bb-0310-9956-ffa450edef68