Removed introspection ( well, moved it to IntrospectionUtil where we can do a better job caching ids, reusing, etc ).
The whole XmlMapper will be probably replaced ( in a future version )
with a code-generation system ( I expect this to happen after jasper34
refactoring is finished, as I'll probably reuse a lot from it ). My estimate
is that 60..80% of the startup time is in this object, and can be eliminated
almost completely. It was a cool idea, but obviously didn't took performance
into consideration ( well, I wrote it - so I'm not offending anyone ).


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283733 13f79535-47bb-0310-9956-ffa450edef68