-  In jk_handler, if a handler is already set ( by a manual configuration ), then deal with that - either call uri_worker_map, or if a single worker is defined use it ( that would avoid an extra call ).
This allows manual mode to be used without double mapping - apache will
do all uri matching once, without beeing duplicated in jk. This is important
for sites with large number of applications, and admins that know apache.
All you have to do is "SetHandler jakarta-servlet" in a <Location> context.

In jk_translate we'll also detect if a handler was explicitely set, but right
now that's useless since we add it as "FIRST" ( that means if both JkMount
and native declarations are used, the mapping will happen anyway ).

I still need to make few improvements, but right now it should be possible to
write apache configs that are completely avoiding duplication of work.


- added a small comment to allow emacs to match the tab style used by the
code author ( and use spaces for new code - since the style is different from
the one used in most java code )
PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/tomcat@295542 13f79535-47bb-0310-9956-ffa450edef68