Log exception messages (with stack traces) when IOException or ClassNotFound- Exception is encountered when unloading sessions to SESSIONS.ser, or reloading them after a restart.
Currently, such errors are silently ignored.  This can cause problems like the
one reported in Bugzilla #2380, where the SESSIONS.ser file was created but
with zero length.  This could happen, for example, if a non-serializable
session attribute was being processed -- the NotSerializableException would
get silently swallowed and no remaining sessions would be saved.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286240 13f79535-47bb-0310-9956-ffa450edef68