Major security addition - DecodeInterceptor is now able to "normalize" a URI, using the same (well tested ) alghoritm that apache does.
This will remove all unsafe path components at the earliest stage,
making sure all components benefit from this ( not only StaticInterceptor
via FileUtils ).

We deal with //, /./, /../, /.., /.

Note that this behavior can be disabled, if you think following
ad-literam the servlet spec is better than security. The servlet spec
requires the user to get the "original" URI - which this fix obviously disables.

On the other side, not doing the normalization opens _huge_ security problems,
and _all_ servlets that are using paths ( a default servlet, etc) will
have to know how to secure the path. Since the container still has problems,
it's very unlikely most user code will be able to do so.

In particular the security matching is greatly affected.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283779 13f79535-47bb-0310-9956-ffa450edef68