Modify the way that the invoker servlet works.
Previously, the invoker servlet would dynamically create a new Wrapper
(and associated servlet mapping) on the first request to a particular
servlet, and then do a RequestDispatcher.forward() to utilize the new
Wrapper immediately.  This causes problems, though, when the invoker is
itself included via a RequestDispatcher -- the semantics of
RequestDispatcher.forward() say that the response is closed after the
forwarded-to servlet returns, so that any output subsequently added by the
including servlet would be ignored.

Now, the invoker servlet will still create a new Wrapper and mapping as it
did before (these are used on all subsequent requests to the same
servlet).  However, instead of calling RequestDispatcher.forward(), a new
servlet instance is allocated, invoked, and deallocated directly.

PR: Bugzilla #1902
Submitted by:	Jason Hunter <jhunter@apache.org>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286275 13f79535-47bb-0310-9956-ffa450edef68