Updated to implement rejection of unsafe escapes, i.e. %25, %2E, %2F, and %5C which are '%','.','/', and '\' respectively.  Note that the escape checking does not occur in the session id that follows ";jsessionid=" if found to be present. This behavior is optional and is controled by the "safe" property, which defaults to true.
Normalization has been moved so that it occurs prior to decoding.  This
means that getRequestURI() will return a normalized URI that remains
encoded as required by the spec.  There is support for this normalization
in RFC1630 (thanks Danny).  The normalization is also optional and
is controlled by the "normalize" property, which defaults to true.

The above rejection and normalization behavior is consistent with what
has been implemented in Tomcat 4.0 and 3.2.3.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283813 13f79535-47bb-0310-9956-ffa450edef68