Fix an as-yet-unfiled bug related to escaping of '\' characters in runtime expressions (which was recently patched in response to Bugzilla #2602. The previous patch caused this code:
  <%= "\u9b5a" %>

to create the (incorrect) Java code:

  out.print("\\9b5a");

in the generated servlet.  Now, backslash escaping occurs *only* outside
the context of a Java string literal, which is assumed to be correctly
escaped according to the requirements of the Java language.

TODO:  Make sure that this does not break the fix for %\> quoting.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286328 13f79535-47bb-0310-9956-ffa450edef68