Fix for the fix.
The original fix allowed modules to be defined in server.xml ( or equivalent ).
If the modules are loaded from modules.xml, the loader will automatically
add the new rules.
If the cached properties files are used - we need to add the rules explicitely.

The "modules" note is also used by ContextXmlReader, to avoid parsing the
same file multiple times.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283869 13f79535-47bb-0310-9956-ffa450edef68