Enhance support for "single sign on" to also support "single sign off" in conformance to the additional spec requirements listed in Servlet Spec 2.3 PFD3, Section 12.5.3.1, page 90.  In particular, under the following circumstances:
- Administrator has configured SingleSignOnValve on a virtual host

- Web applications on this virtual host are configured for form-based
  login

- User has logged in to one of the apps (which also grants them access
  to the protected areas of the other apps, because of the single sign
  on support)

- A session in one of the logged-on-to apps is invalidated or times out

then, all sessions in all apps to which the user has been logged on will
be invalidated.  The next request from this user for a protected resource
in any of the apps will cause the user to be reauthenticated.

Rationale:  Consider a Yahoo-like portal server, with multiple
applications (such as mail and chat) implemented as separate web apps
under management of the single sign on facility.  If the user accesses
both mail and chat, and then logs out of mail, he or she will expect the
login to the chat application to be terminated as well.  Otherwise, the
next person to use this browser (say, at an Internet cafe) would have
access to the previous user's session in the chat application -- and, by
virtue of single sign on, access to all their other applications as well.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286366 13f79535-47bb-0310-9956-ffa450edef68