Session/StandardSession:  Implement a new "notes" API that is functionally equivalent to session attributes, but available only to internal Catalina components (not applications).  This can be used by event listeners to decorate sessions with arbitrary sets of objects.
StandardSession:  Implement property change notification on the "authType"
and "principal" properties, so that event listeners can track changes to
these properties (used to cache authentication information and avoid hving
to call Realm.authenticate() on every request).

FormAuthenticator:  Use the new "notes" mechanism, rather than session
attributes, to store the objects related to form-based login processing.
Also, store additional information so that form based login continues to
work even if the "cache" property is set to false (which will cause a call
to Realm.authenticate() on every request).

This is the third of three patches necessary to fix BugTraq #4485977.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286374 13f79535-47bb-0310-9956-ffa450edef68