Omnibus changes to clean up interactions with J2EE RI's security model.
Request/RequestBase - Implement a "notes" capability like that recently added
for sessions.

HttpRequest/HttpRequestBase - Remove the "ssoId" property, which was used to
track the "single sign on" identity (if any) for this request.  This value
is now passed as a note.

InstanceEvent/ApplicationDispatcher - Send "before dispatch" and "after
dispatch" events around the service() call of a dispatched-to servlet, so
"run as" identity lookups look in the correct place.

org.apache.catalina.authenticator.* - Several changes
- Pass the username and password used to authenticate the user (if any)
  as request notes so that the EJB layer can access them
- Register the username and password with SingleSignOnValve as needed


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286378 13f79535-47bb-0310-9956-ffa450edef68