Added explicit setContextManager() ( to prevent eventual NPE ).
Empty init/destroy ( most internal handlers don't need that, but it's better to
avoid future class loader problems )

Invoke() is documented as the prefered method for handlers, but it was broken. Now
it works.

Added the simpler get/setNote with "real" name instead of name id. Slower but simpler,
shouldn't be used in the critical path.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283885 13f79535-47bb-0310-9956-ffa450edef68