Few fixes here. The changes in startup showed few obscure bugs on log initialization.
Note that LogManager is used as a guard on all sensitive accesses to Log. Now adding a
channel will 'fix' all logs that were created earlier for the channel ( with default
logger )

This is important for loggers that are created before LogSetter module gets a chance
to initialize the logging subsystem. In normal operation we don't do any logging before
that anyway, but for debugging it's usefull ( and makes the package more usefull for
use outside tomcat )


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283960 13f79535-47bb-0310-9956-ffa450edef68