Fixes in sandboxing.
Add a message advising to set -Djava.security.policy. On some VMs it is possible
to set it later, but in some it isn't - the right way to run the sandbox is to
make sure a policy is defined.

PolicyInterceptor will enable the sandbox - it is not required if you embed
tomcat and have a different mechanism to set sandboxing. As with other modules,
it just provide a default and/or template for more advanced modules.

Also added a Policy.refresh(), few more log statements.

Updated the default permissions to include read for lib/common and read/apps
( but not read/container, of course ).

Added "getClassLoader" permission by default, it's needed by jaxp, jaxm, etc.

Added a "sandbox" option on PolicyLoader to force the use of the sandbox,
the default is to use it only if a "sandbox" property is set on context manger.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@283965 13f79535-47bb-0310-9956-ffa450edef68