Add a replacement for the java.net.URL class that does the parsing and relative-URL resolution performed by java.net.URL, but does not suffer from the limitation that it only supports protocols for which a URLStreamHandler is installed.  In particular, this was preventing a Tomcat 4 install from successfully performing a response.sendRedirect() to an SSL-enabled URL, unless the redirecting Tomcat itself had JSSE installed.
Also added extensives JUnit tests to ensure that parsing and relative
resolution works as much as possible like java.net.URL does it.  There are
three cases where the answers are different, that are commented out.
These are really edge cases that won't matter in real life, but they
should eventually be corrected.

NOTE:  The parsing and path normalization code is horrendously inefficient
at the moment.  I focused first on getting the behavior right, and
creating a bunch of unit tests so that we can make sure we don't introduce
regressions as the string processing is optimized.

PR: Bugzilla #3139
Submitted by:	Shawn Bayern <bayern@essentially.net>


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286441 13f79535-47bb-0310-9956-ffa450edef68