Enhance Catalina startup/shutdown processing to support a new (optional) environment variable (CATALINA_BASE), and corresponding system property (catalina.base).  If not set, these values default to corresponding values for CATALINA_HOME and catalina.home, respectively.
When you set CATALINA_BASE to point to a directory other than the one
pointed to by CATALINA_HOME, you are changing the *default* base from
which all references to files in the "conf", "logs", "webapps", and "work"
directories are resolved.  This allows you to share the static portion of
a Tomcat 4 installation (read-only) among multiple uses of that
installation on the same server.

TODO:  A pass through the docs to update them with this new information is
coming up shortly.

Submitted by:  Too numerous to count :-)


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286450 13f79535-47bb-0310-9956-ffa450edef68