Cleaning up old TODO items ...
Split out the built-in servlets that are part of Catalina into their own
JAR files in $CATALINA_HOME/server/lib so that you can grant individual
security manager permissions to each servlet (if you want to), or -- for
paranoid sysadmins -- remove them entirely if you do not need that
particular servlet's functionality in your environment.

The "servlets-common.jar" file contains things that are common to all of
the servlets (so it needs to stay in "/server/lib" if *any* of them are
used).  Each of the other "servlets-*.jar" files contains only the code
for the corresponding built-in servlet.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286479 13f79535-47bb-0310-9956-ffa450edef68