Because JSP source and class files are now stored in a directory tree that mirrors the context-relative path, there is no need to include the directory path in the generated filenames as well.
This is based on the patch submitted by Kin-Man Chung
<kin-man.chung@sun.com> but modified to fix an additional bug -- the
previous algorithm would only work when JSP pages were mapped to "*.jsp",
but would fail when a different servlet mapping was mapped to the JSP
servlet.  In addition, name mangling is still necessary to convert
characters that are not legal Java identifiers.

PR: Bugzilla #2917
Submitted by:	aklimuk@iba.by


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286501 13f79535-47bb-0310-9956-ffa450edef68