Modify the org.apache.catalina.startup.Embedded class to *require* that the "catalina.home" system property be set in order to function.  Many internal Catalina operations depend on this.
Modify the org.apache.catalina.startup.Embedded class to copy the value of
the "catalina.home" system property to the "catalina.base" system property
if the latter is not set.  Many internal Catalina operations depend on
this.

Add a backwards-compatibility patch so that the most recent Tomcat 4.0
code will still run with the J2EE 1.3 Reference Implementation.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286533 13f79535-47bb-0310-9956-ffa450edef68