Bug 3233, make error handling compatible with servlet2.3 by using the clarification from 2.3 spec ( it wasn't technically a bug, since 2.2 leaves this part unspecified ). Thanks ps@psncc.at (Peter Stamfest) for the patch.
I did a small change for 3.3, since the ErrorHandler sits at a lower level and doesn't
depend on the facade ( it is part of the set of core modules ). Since error handling is
so dependent on the servlet api, in future we could move part of it in the facade, but
for now it should be ok.

Submitted by:	Peter Stamfest


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284135 13f79535-47bb-0310-9956-ffa450edef68