port sam's tyrex dependency patch back to the branch
fix things so that you don't have to hand edit this file in order to build
tomcat. the problem is that the JSSE stuff was commented out and things
wouldn't compile because of that. if you wanted to compile, you had to
hand edit this file and uncomment things. the real issue is that the JSSE
stuff is only a problem when you want to make a distribution...so don't
allow people to do that.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@286615 13f79535-47bb-0310-9956-ffa450edef68