Usability fix: if a webapplication doesn't include a parser add the default one.
This behavior can be turned off, and can be configured with an alternate
parser ( by specifying the dir and the jars ). It'll do nothing if the
application includes jaxp.

We still use the standard loading mechanism ( using delegation to parent first),
while providing a parser without preventing the webapp from defining it's own.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284222 13f79535-47bb-0310-9956-ffa450edef68