Another small usability improvement.
Modules ( == trusted contexts, including tomcat-add on services ) can now add classes to
the container and common loader. That eliminate a lot of configuration effort and provides a
very simple mechanism to add libs, using regular war files.

Note that nothing is changed in tomcat operation if no module is included ( the current
case ), it should operate exactly as before.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284225 13f79535-47bb-0310-9956-ffa450edef68