Fix Bug #3944 request.getRemoteAddr() alays returning 127.0.0.1 Bug report from Alessandro Polverini
Since the base Request object set remoteAddr and remoteHost to defaults,
the checks around setting them off of the real request always succeed so the
real values are never used.  Added calls to recycle these in the constructor.
This still allows for delayed setting, since the calls to set the values can be
expensive.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284242 13f79535-47bb-0310-9956-ffa450edef68