getRequestURI returns an encoded request. Fix is slightly modified from what Bill Barker sent, I was thinking to do some optimizations ( avoid use of Writer, and do the encoding directly in the buffer), but given that this happens only when it's needed - it should be ok. Probably for 3.3.1 we can do this ( and the date formating which is another hotspot ).
Note that we do not encode contextPath. This is against the spec, but I think
we should leave it this way:
- that's how it was in 3.1, 3.2, etc
- 4.0 is also broken ( well, it returns 404 if I try any encoded char in the context
path )
- it can raise a large number of bugs, if users expect the contextPath to be unique
( with encoding - there are many ways you can encode a string, without
even considering charsets ).
- it would be very hard to write an app that is consistent between 3.3 and 3.2
( which most people use ).

I don't know how other containers are doing this, but I suspect most will
return the context path decoded - if you can find one that is not decoding the
CP - we can reconsider this.


git-svn-id: https://svn.apache.org/repos/asf/tomcat@284244 13f79535-47bb-0310-9956-ffa450edef68